deleted empty test classes and centralised GenAI prompt construction
