+++ windows/ManuscriptaTeacherApp/Main/Services/GenAI/ContentModificationService.cs
@@ -54,7 +54,7 @@ public class ContentModificationService
         }
 
         // §3C(2)(b): Construct prompt
-        var prompt = ConstructModificationPrompt(selectedContent, instruction, relevantChunks);
+        var prompt = GenAIPromptBuilder.BuildModificationPrompt(selectedContent, instruction, relevantChunks);
 
         // §3C(2)(c): Invoke model
         var modifiedContent = await _ollamaClient.GenerateChatCompletionAsync(ModificationModel, prompt);
@@ -65,43 +65,4 @@ public class ContentModificationService
         return result;
     }
 
-    /// <summary>
-    /// Constructs the modification prompt.
-    /// See GenAISpec.md §3C(2)(b) and Material Encoding Specification §4.
-    /// </summary>
-    private string ConstructModificationPrompt(
-        string selectedContent,
-        string instruction,
-        List<string> relevantChunks)
-    {
-        var contextSection = relevantChunks.Count > 0
-            ? $"Relevant context from source documents:\n{string.Join("\n\n", relevantChunks)}\n\n"
-            : "";
-
-        return $@"{contextSection}Modify the following content according to the instruction provided.
-
-Original content:
-{selectedContent}
-
-Instruction: {instruction}
-
-Return the modified content following these Markdown syntax requirements:
-- Headers shall use # for Level 1, ## for Level 2, ### for Level 3 (do not exceed Level 3)
-- Text shall be rendered in bold using **text** or __text__ syntax
-- Text shall be rendered in italic using *text* or _text_ syntax
-- Unordered lists shall be specified using - item or * item syntax
-- Ordered lists shall be specified using 1. item syntax
-- Tables use GitHub Flavored Markdown syntax with pipes (|) and dashes (---)
-- Code blocks shall be specified using triple backticks with optional language identifier
-- Blockquotes shall be specified using the > prefix
-- Horizontal rules shall be specified using three or more hyphens on a line
-- LaTeX notation shall be delimited by single dollar signs ($...$) for inline or double dollar signs ($$...$$) for blocks
-- Images shall be embedded using ![alt text](/attachments/{{attachment-uuid}})
-- PDF documents shall be embedded using the pdf marker: !!! pdf id=""{{pdf-attachment-uuid}}""
-- Text shall be centred using the center marker: !!! center
-- Questions shall be referenced using the question marker: !!! question id=""{{question-uuid}}""
-- Preserve existing formatting and structure where appropriate
-
-Modified content:";
-    }
 }
