--- windows/ManuscriptaTeacherApp/MainTests/ServicesTests/GenAI/ContentModificationServiceTests.cs
@@ -1,37 +1,0 @@
-using System.Collections.Generic;
-using System.Reflection;
-using Main.Services.GenAI;
-using Xunit;
-
-namespace MainTests.ServicesTests.GenAI;
-
-/// <summary>
-/// Spec coverage: GenAISpec Section 3C (Content Modification).
-/// See docs/specifications/GenAISpec.md.
-/// </summary>
-public class ContentModificationServiceTests
-{
-    /// <summary>
-    /// Spec coverage: GenAISpec Section 3C(2)(b)(iii) (optional context inclusion).
-    /// See docs/specifications/GenAISpec.md.
-    /// </summary>
-    [Fact(Skip = "ConstructModificationPrompt is private; prompt behavior should be tested via a public/internal API instead of reflection.")]  
-    public void ConstructModificationPrompt_NoContext_DoesNotIncludeContextSection()  
-    {  
-        // This test previously invoked the private ConstructModificationPrompt method via reflection.  
-        // To avoid brittle tests that depend on private implementation details, this test is skipped.  
-        // Prompt construction should instead be verified indirectly through a public or internal API.  
-    }  
-
-    /// <summary>
-    /// Spec coverage: GenAISpec Section 3C(2)(b)(iii) (context included when provided).
-    /// See docs/specifications/GenAISpec.md.
-    /// </summary>
-    [Fact(Skip = "ConstructModificationPrompt is private; prompt behavior should be tested via a public/internal API instead of reflection.")]  
-    public void ConstructModificationPrompt_WithContext_IncludesContextSection()  
-    {  
-        // This test previously invoked the private ConstructModificationPrompt method via reflection.  
-        // To avoid brittle tests that depend on private implementation details, this test is skipped.  
-        // Prompt construction should instead be verified indirectly through a public or internal API. 
-    }
-}
