--- windows/ManuscriptaTeacherApp/MainTests/ServicesTests/GenAI/MaterialGenerationServiceTests.cs
@@ -1,35 +1,0 @@
-using System.Collections.Generic;
-using System.Reflection;
-using Main.Services.GenAI;
-using Xunit;
-
-namespace MainTests.ServicesTests.GenAI;
-
-/// <summary>
-/// Spec coverage: GenAISpec Section 3B (Material Generation prompts).
-/// See docs/specifications/GenAISpec.md.
-/// </summary>
-public class MaterialGenerationServiceTests
-{
-    /// <summary>
-    /// Spec coverage: GenAISpec Section 3B(3)(c)(v) (exclude Questions section for readings).
-    /// See docs/specifications/GenAISpec.md.
-    /// </summary>
-    [Fact]
-    public void ConstructGenerationPrompt_ReadingMaterial_ExcludesQuestionSyntax()
-    {
-        // This test previously invoked the private ConstructGenerationPrompt method via reflection.
-        // To avoid brittle tests that depend on private implementation details, this test is skipped.
-    }
-
-    /// <summary>
-    /// Spec coverage: GenAISpec Section 3B(3)(c)(iv)-(v) (include question-draft syntax for worksheets).
-    /// See docs/specifications/GenAISpec.md.
-    /// </summary>
-    [Fact]
-    public void ConstructGenerationPrompt_WorksheetMaterial_IncludesQuestionSyntax()
-    {
-        // This test previously invoked the private ConstructGenerationPrompt method via reflection.
-        // To avoid brittle tests that depend on private implementation details, this test is skipped.
-    }
-}
